weave_introduction_handson/pyproject.toml
[project]
name = "weave-introduction-handson"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "google-genai>=1.42.0",
    "google-generativeai>=0.8.6",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.5.0",
    "jupyter>=1.1.1",
    "nest-asyncio>=1.6.0",
    "numpy>=2.3.3",
    "openai>=1.76.2",
    "openai-agents>=0.0.3",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "reportlab>=4.4.9",
    "wandb>=0.19.10",
    "weave>=0.52.22",
]
